#S07P31B102-234 feat.없는 데이터 에러 처리 구현중

+++ Back/selly-user-service/src/main/java/com/b102/sellyuserservice/vo/NftPieceRequest.java
@@ -9,11 +9,11 @@ import lombok.NoArgsConstructor;
 public class NftPieceRequest {
   private Long articleId;
   private Long userId;
-  private Long nftPieceCnt;
+  private Integer nftPieceCnt;
   private double avgPrice;
 
   @Builder
-  public NftPieceRequest(Long articleId, Long userId, Long nftPieceCnt, double avgPrice) {
+  public NftPieceRequest(Long articleId, Long userId, Integer nftPieceCnt, double avgPrice) {
     this.articleId = articleId;
     this.userId = userId;
     this.nftPieceCnt = nftPieceCnt;
